New Dynamic Dialogue System (ADDS)
The last week has been spent extracting the dialogue system used in PhilTroy, expanding and modularizing it (for easier reutilization in other games) and putting it back into PhilTroy. The end result is called "Ape´s Dynamic Dialogue System" (or ADDS for short) This is an important step, which at first glance doesn´t seem all that significant, but the new features of this system are:

- NPCs remember what they´ve talked about before
- You can flip through replies and select reply with the keyboard
- You can alter game variables through dialogue choices
- Dialogue paging allows for longer dialogues
- all dialogue functionality moved to its´own script
- code was cleaned up for better readability

You can check the junkheap repository > extended dialogue system for full feature list.

ADDS will be made into a plugin, but that´s for later.